refactor(ai-panel): 抽出 AIPanelContent.vue，AIPanelView 暫為過渡外殼
- 將 AIPanelView.vue 的 SEO 生成區塊與相關邏輯搬入新元件 AIPanelContent.vue，
  改用 Task 1 拆分出的 useSeoResultStore 取代已刪除的 aiPanel store
- AIPanelView.vue 保留外框（border/resize-handle/panel header/data-testid），
  改為單純渲染 <AIPanelContent>，作為過渡態供後續整批替換
- 順手修正 App.vue 對已刪除 aiPanel store 的殘留 import，
  改用 Task 1 已建立的 useInspectorStore（isOpen/toggle 介面相同），
  否則 renderer build 會直接失敗，無法驗證任何 E2E
