+++ src/components/AIPanelView.vue
@@ -9,101 +9,23 @@
         <Sparkles :size="16" class="text-primary" />
         <span class="font-semibold text-sm">AI 助手</span>
       </div>
-      <button class="btn btn-ghost btn-xs btn-square" data-testid="ai-panel-close-button" @click="aiPanelStore.close()">
+      <button class="btn btn-ghost btn-xs btn-square" data-testid="ai-panel-close-button" @click="inspectorStore.close()">
         <X :size="14" />
       </button>
     </div>
 
-    <!-- No Article State -->
-    <div v-if="!article" class="flex-1 flex items-center justify-center p-6">
-      <p class="text-sm text-base-content/50 text-center">請先選擇一篇文章</p>
-    </div>
-
-    <!-- Panel Content -->
-    <div v-else class="flex-1 overflow-y-auto">
-
-      <!-- SEO 生成 -->
-      <div class="border-b border-base-300">
-        <button
-          class="w-full flex items-center justify-between px-4 py-2 bg-base-200 hover:bg-base-300 transition-colors text-sm font-medium"
-          @click="seoExpanded = !seoExpanded"
-        >
-          <div class="flex items-center gap-2">
-            <component :is="seoExpanded ? ChevronDown : ChevronRight" :size="14" />
-            <span>SEO 生成</span>
-          </div>
-        </button>
-
-        <div v-if="seoExpanded" class="p-4 space-y-3">
-          <!-- No API Key -->
-          <div v-if="!hasApiKey" class="text-center space-y-2">
-            <p class="text-xs text-base-content/60">需要 API Key 才能使用 AI 功能</p>
-            <button class="btn btn-xs btn-outline" @click="$emit('open-settings', 'ai')">⚙ 前往設定</button>
-          </div>
-
-          <div v-else class="space-y-3">
-            <button
-              class="btn btn-sm btn-primary w-full"
-              :disabled="seoStore.isGenerating"
-              @click="handleGenerateSEO"
-            >
-              <span v-if="!seoStore.isGenerating" class="flex items-center gap-1">
-                <Sparkles :size="14" /> 生成 SEO
-              </span>
-              <span v-else class="loading loading-spinner loading-xs"></span>
-            </button>
-
-            <!-- Error -->
-            <div v-if="aiPanelStore.seoError" class="alert alert-error p-2">
-              <span class="text-xs">{{ aiPanelStore.seoError }}</span>
-            </div>
-
-            <!-- Result -->
-            <div v-if="aiPanelStore.seoResult" class="space-y-2">
-              <div class="rounded-lg border border-base-300 overflow-hidden text-xs">
-                <div class="p-3 space-y-2">
-                  <div>
-                    <span class="text-base-content/50 text-xs uppercase tracking-wide">Slug</span>
-                    <p class="font-mono mt-0.5 break-all">{{ aiPanelStore.seoResult.slug }}</p>
-                  </div>
-                  <div>
-                    <span class="text-base-content/50 text-xs uppercase tracking-wide">Description</span>
-                    <p class="mt-0.5">{{ aiPanelStore.seoResult.metaDescription }}</p>
-                  </div>
-                  <div>
-                    <span class="text-base-content/50 text-xs uppercase tracking-wide">Keywords</span>
-                    <div class="flex flex-wrap gap-1 mt-0.5">
-                      <span
-                        v-for="kw in aiPanelStore.seoResult.keywords"
-                        :key="kw"
-                        class="badge badge-outline badge-sm"
-                      >{{ kw }}</span>
-                    </div>
-                  </div>
-                </div>
-              </div>
-              <div class="flex gap-2">
-                <button class="btn btn-xs btn-primary flex-1" @click="handleApplySEO">套用到文章</button>
-                <button class="btn btn-xs btn-ghost" @click="aiPanelStore.clearSEO()">清除</button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-
-    </div>
+    <AIPanelContent :article="article" @open-settings="$emit('open-settings', $event)" />
   </div>
 </template>
 
 <script setup lang="ts">
 import { ref, onMounted, onUnmounted } from "vue"
-import { Sparkles, X, ChevronDown, ChevronRight } from "@lucide/vue"
-import { useAIPanelStore } from "@/stores/aiPanel"
-import { useSeoStore } from "@/stores/seo"
-import { useArticleStore } from "@/stores/article"
+import { Sparkles, X } from "@lucide/vue"
+import { useInspectorStore } from "@/stores/inspector"
+import AIPanelContent from "@/components/AIPanelContent.vue"
 import type { Article } from "@/types"
 
-const props = defineProps<{
+defineProps<{
   article: Article | null
 }>()
 
@@ -111,12 +33,7 @@ defineEmits<{
   "open-settings": [tab?: string]
 }>()
 
-const aiPanelStore = useAIPanelStore()
-const seoStore = useSeoStore()
-const articleStore = useArticleStore()
-
-const seoExpanded = ref(true)
-const hasApiKey = ref(false)
+const inspectorStore = useInspectorStore()
 
 // Resize
 const MIN_WIDTH = 240
@@ -147,7 +64,7 @@ function stopResize() {
   }
 }
 
-onMounted(async () => {
+onMounted(() => {
   const savedWidth = localStorage.getItem(STORAGE_KEY)
   if (savedWidth) {
     const parsed = Number.parseInt(savedWidth, 10)
@@ -157,30 +74,12 @@ onMounted(async () => {
   }
   document.addEventListener("mousemove", handleMouseMove)
   document.addEventListener("mouseup", stopResize)
-
-  hasApiKey.value = await seoStore.hasApiKey()
 })
 
 onUnmounted(() => {
   document.removeEventListener("mousemove", handleMouseMove)
   document.removeEventListener("mouseup", stopResize)
 })
-
-async function handleGenerateSEO() {
-  const article = props.article
-  if (!article) { return }
-  await aiPanelStore.generateSEO(article)
-  hasApiKey.value = await seoStore.hasApiKey()
-}
-
-function handleApplySEO() {
-  const article = props.article
-  if (!article) { return }
-  const updated = aiPanelStore.applySEOResult(article)
-  if (updated) {
-    articleStore.updateArticleInMemory(updated)
-  }
-}
 </script>
 
 <style scoped>
